Backend/requirements.txt
flask
flask-cors
gunicorn
python-dotenv
google-genai
